finish up part 2
